requirements/requirements_test.txt
pytest==3.4.0
pytest-django==3.1.2
factory_boy==2.10.0
pytest_factoryboy==2.0.1
Faker==0.8.10
pytest-cov==2.5.1